refactor(finance): split service by responsibility
