+++ src/lifeos_cli/db/services/finance/__init__.py
@@ -1,0 +1,130 @@
+"""Public finance service facade.
+
+The facade preserves the historical ``lifeos_cli.db.services.finance`` import
+surface while keeping implementation modules organized by responsibility.
+"""
+
+from ._core import (
+    DEFAULT_FINANCE_ASSET_DECIMAL_PLACES,
+    DEFAULT_FINANCE_ASSETS,
+    DEFAULT_FINANCE_CURRENCY,
+    FINANCE_AMOUNT_QUANT,
+    FINANCE_RATE_QUANT,
+    MAX_FINANCE_ASSET_DECIMAL_PLACES,
+    FinanceAssetAlreadyExistsError,
+    FinanceAssetNotFoundError,
+    FinanceRateSnapshotEntryInput,
+    FinanceRateSnapshotNotFoundError,
+    FinanceSnapshotEntryInput,
+    FinanceTreeAlreadyExistsError,
+    FinanceTreeNodeAlreadyExistsError,
+    FinanceTreeNodeNotFoundError,
+    FinanceTreeNotFoundError,
+    FinanceValidationError,
+    RateResolution,
+    decimal_quant_for_places,
+    format_asset_amount,
+    format_decimal_places,
+    normalize_currency_code,
+    validate_asset_decimal_places,
+    validate_asset_name,
+    validate_node_name,
+    validate_snapshot_title,
+    validate_tree_name,
+)
+from .assets import (
+    count_finance_assets,
+    create_finance_asset,
+    delete_finance_asset,
+    ensure_default_finance_assets,
+    list_finance_assets,
+    update_finance_asset,
+)
+from .lifecycle import delete_finance_rate_snapshot, update_finance_rate_snapshot
+from .rates import (
+    count_finance_rate_snapshots,
+    create_finance_rate_snapshot,
+    get_finance_rate_snapshot,
+    list_finance_rate_snapshots,
+)
+from .snapshots import (
+    count_finance_snapshots,
+    create_finance_snapshot,
+    delete_finance_snapshot,
+    get_finance_snapshot,
+    list_finance_snapshots,
+    update_finance_snapshot,
+)
+from .trees import (
+    count_finance_trees,
+    create_finance_node,
+    create_finance_tree,
+    delete_finance_node,
+    delete_finance_tree,
+    ensure_default_finance_tree,
+    get_finance_tree,
+    get_finance_tree_with_nodes,
+    list_finance_nodes,
+    list_finance_trees,
+    update_finance_node,
+    update_finance_tree,
+)
+
+__all__ = [
+    "DEFAULT_FINANCE_ASSET_DECIMAL_PLACES",
+    "DEFAULT_FINANCE_ASSETS",
+    "DEFAULT_FINANCE_CURRENCY",
+    "FINANCE_AMOUNT_QUANT",
+    "FINANCE_RATE_QUANT",
+    "MAX_FINANCE_ASSET_DECIMAL_PLACES",
+    "FinanceAssetAlreadyExistsError",
+    "FinanceAssetNotFoundError",
+    "FinanceRateSnapshotEntryInput",
+    "FinanceRateSnapshotNotFoundError",
+    "FinanceSnapshotEntryInput",
+    "FinanceTreeAlreadyExistsError",
+    "FinanceTreeNodeAlreadyExistsError",
+    "FinanceTreeNodeNotFoundError",
+    "FinanceTreeNotFoundError",
+    "FinanceValidationError",
+    "RateResolution",
+    "decimal_quant_for_places",
+    "format_asset_amount",
+    "format_decimal_places",
+    "normalize_currency_code",
+    "validate_asset_decimal_places",
+    "validate_asset_name",
+    "validate_node_name",
+    "validate_snapshot_title",
+    "validate_tree_name",
+    "count_finance_assets",
+    "create_finance_asset",
+    "delete_finance_asset",
+    "ensure_default_finance_assets",
+    "list_finance_assets",
+    "update_finance_asset",
+    "count_finance_rate_snapshots",
+    "create_finance_rate_snapshot",
+    "delete_finance_rate_snapshot",
+    "get_finance_rate_snapshot",
+    "list_finance_rate_snapshots",
+    "update_finance_rate_snapshot",
+    "count_finance_snapshots",
+    "create_finance_snapshot",
+    "delete_finance_snapshot",
+    "get_finance_snapshot",
+    "list_finance_snapshots",
+    "update_finance_snapshot",
+    "count_finance_trees",
+    "create_finance_node",
+    "create_finance_tree",
+    "delete_finance_node",
+    "delete_finance_tree",
+    "ensure_default_finance_tree",
+    "get_finance_tree",
+    "get_finance_tree_with_nodes",
+    "list_finance_nodes",
+    "list_finance_trees",
+    "update_finance_node",
+    "update_finance_tree",
+]
